# crypto: optimize Keccak*/SHA3 calls by re-using hashes