--- a/crypto/crypto_test.go
+++ b/crypto/crypto_test.go
@@ -61,7 +61,8 @@ func TestToECDSAErrors(t *testing.T) {
 func BenchmarkSha3(b *testing.B) {
 	a := []byte("hello world")
 	for i := 0; i < b.N; i++ {
-		Keccak256(a)
+		h := Keccak256(a)
+		b.SetBytes(int64(len(h)))
 	}
 }