@holiman 2022-04-28
In theory you have a point, however, the `crypto.Keccak256Hash` is seldom used in 'hot paths' -- and the hot paths already most often use their own hashers (with reset). 
A quick check: the 'hottest' path that I can see is when a new contract is created, and we hash the code. But still, that only happens maybe a couple of times per block, so it's not particilarly hot anyway.