@fjl 2022-05-05
The benchmark results look impressive, but are also meaningless. It would be better to benchmark some code that actually uses these functions.

As @holiman says, most performance-critical code paths reuse their hash instance already. To improve `GetCommittedState`, we could use/maintain a hash instance as part of `StateDB`.