@odeke-em 2022-05-05
> The benchmark results look impressive, but are also meaningless. It would be better to benchmark some code that actually uses these functions.
> 
> 
> 
> As @holiman says, most performance-critical code paths reuse their hash instance already.
> 
> 
> 
> To improve `GetCommittedState`, we could use/maintain a hash instance as part of `StateDB`.

Thanks for the input @fjl and @holiman!
I don't seem to understand the attitude that clear improvements are "meaningless" and should only be to the super critical Ethereum code paths yet a whole lot of code is in packages with public APIs that some of my customers use (which is why I spent time to profile and then send these PRs)
I am working my way through various improvements that use Ethereum, as my customers' needs expose, but it surprises me that clear improvements firstly require one to grok through parts of the Ethereum code base that my customers don't use directly.
Is there some sort of policy on what packages are cared about for performance improvements? Should I perhaps instead be filing issues then letting core maintainers work on them? I ask because I find these things while on time constrained audits and after profiling, so perhaps would be better to report them then the maintainers with much more context can sort them? Thank you.