@fjl 2022-05-05
Maybe even more context:

For most code in core/, crypto/, eth/ etc. areas of go-ethereum, we only care about performance in the context of our project "geth", and basically assume by default that all PRs making changes there have the ultimate goal of improving "geth" performance.

This is why we may occasionally start discussing alternative changes in other parts of go-ethereum on PRs. It doesn't mean you need to go ahead and make those changes. We're just thinking out loud.